Check upload.

+++ resources/virtocommerce/DocsUpdate.ps1
@@ -2,6 +2,7 @@ $ErrorActionPreference = 'Continue'
 
 $DestContentPath = ""
 $ZipFile = "${env:ArtifactPath}"
+Write-Host "ZipFile: $ZipFile"
 
 $ApplicationID ="${env:AzureAppID}"
 $APIKey = ConvertTo-SecureString "${env:AzureAPIKey}" -AsPlainText -Force
@@ -15,7 +16,7 @@ Select-AzureRmSubscription -SubscriptionId $SubscriptionID
 $DestResourceGroupName = "${env:AzureResourceGroupNameProd}"
 $DestWebAppName = "${env:AzureWebAppNameProd}"
 
-Write-Host "Stop WebApp $DestWebAppName"
+# Write-Host "Stop WebApp $DestWebAppName"
 # Stop-AzureRmWebApp -ResourceGroupName $DestResourceGroupName -Name $DestWebAppName
 
 # Start-Sleep -s 35
@@ -31,14 +32,14 @@ $contentPath = $DestContentPath
 $msdeploy = "${env:MSDEPLOY_DIR}\msdeploy.exe"
 $sourcewebapp_msdeployUrl = "https://${DestWebAppName}.scm.azurewebsites.net/msdeploy.axd?site=${DestWebAppName}"
 & $msdeploy -verb:sync -dest:contentPath="D:\home\site\wwwdocs\latest\$contentPath",computerName=$sourcewebapp_msdeployUrl,publishSettings=$tmpPublishProfile -source:package=$ZipFile
-# if($LASTEXITCODE -ne 0)
-# {
-#     exit 1
-# }
+if($LASTEXITCODE -ne 0)
+{
+    exit 1
+}
 
 # Start-Sleep -s 5
 
-Write-Host "Start WebApp $DestWebAppName"
+# Write-Host "Start WebApp $DestWebAppName"
 # Start-AzureRmWebApp -ResourceGroupName $DestResourceGroupName -Name $DestWebAppName
 
 Remove-Item $tmpPublishProfile -Force
